Changes setup option --freedom to manditory.
Reason:

Nopony should be by mistake install non-free ponies thinking that
that are free because the license of ponysay itself is free.
But nopony should either just make in installation and miss out
one all the ponies just because the he had not read the information
on how to install the package.